calendar: phase 3a extract shared parser scalar helpers
First, low-risk stage of the parser de-duplication. Move the format-agnostic
string helpers shared verbatim by the XML and JSON document parsers
(ParseEnum/ParseNullableEnum, ParseWorkingWeek, ParseMonth, ParseTriggerMonth,
ParseFixedMonth, the month-name table) into a shared internal
NotableDateDocumentParseHelpers, referenced via 'using static'. Diagnostics
codes/messages are unchanged.

Calendar regression (2132) green; XML/JSON parser-equivalence tests intact.
