+++ Bodu.Globalization.Calendar/src/Globalization.Calendar/NotableDateJsonDocumentParser.cs
@@ -7,6 +7,7 @@
 using System.Globalization;
 using System.Text.Json;
 using Bodu.Extensions;
+using static Bodu.Globalization.Calendar.NotableDateDocumentParseHelpers;
 
 namespace Bodu.Globalization.Calendar;
 
@@ -17,15 +18,6 @@ namespace Bodu.Globalization.Calendar;
 /// </summary>
 internal static class NotableDateJsonDocumentParser
 {
-    /// <summary>
-    /// The full English month names, indexed so that January is at index zero.
-    /// </summary>
-    private static readonly string[] s_monthNames =
-    [
-        "January", "February", "March", "April", "May", "June",
-        "July", "August", "September", "October", "November", "December",
-    ];
-
     /// <summary>
     /// Parses a notable-date document from its JSON content.
     /// </summary>
@@ -123,15 +115,6 @@ internal static class NotableDateJsonDocumentParser
     }
 
     /// <summary>
-    /// Parses a Sunday-first seven-character working-week pattern, falling back to the default working week when the
-    /// value is absent or blank.
-    /// </summary>
-    /// <param name="value">The working-week pattern, or <see langword="null" />.</param>
-    /// <returns>The parsed <see cref="WeekPattern" />, or <see langword="null" /> when unspecified.</returns>
-    private static WeekPattern? ParseWorkingWeek(string? value) =>
-        string.IsNullOrWhiteSpace(value) ? null : WeekPattern.Parse(value);
-
-    /// <summary>
     /// Parses the reusable adjustment policies declared by the resource.
     /// </summary>
     /// <param name="element">The <c>adjustmentPolicies</c> array, or <see langword="null" />.</param>
@@ -195,23 +178,6 @@ internal static class NotableDateJsonDocumentParser
             result[property.Name] = property.Value.ValueKind == JsonValueKind.String ? property.Value.GetString() ?? string.Empty : property.Value.ToString();
 
         return result.Count > 0 ? result : null;
-    }
-
-    /// <summary>
-    /// Parses a trigger comparison month expressed as a full English month name or an integer between 1 and 12.
-    /// </summary>
-    /// <param name="value">The raw month value, or <see langword="null" />.</param>
-    /// <returns>The one-based month, or <see langword="null" /> when absent or invalid.</returns>
-    private static int? ParseTriggerMonth(string? value)
-    {
-        if (value is null)
-            return null;
-
-        if (int.TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out var numeric) && numeric is >= 1 and <= 12)
-            return numeric;
-
-        var index = Array.FindIndex(s_monthNames, n => string.Equals(n, value, StringComparison.OrdinalIgnoreCase));
-        return index >= 0 ? index + 1 : null;
     }
 
     /// <summary>
@@ -462,85 +428,6 @@ internal static class NotableDateJsonDocumentParser
     }
 
     /// <summary>
-    /// Parses a month value expressed as a full English month name or an integer between 1 and 12.
-    /// </summary>
-    /// <param name="value">The raw month value.</param>
-    /// <param name="notableDateId">The identifier of the owning concept, used in diagnostics.</param>
-    /// <param name="ruleId">The identifier of the owning rule, used in diagnostics.</param>
-    /// <param name="diagnostics">The collection that receives semantic diagnostics.</param>
-    /// <returns>The one-based month, or 1 when the value is invalid.</returns>
-    private static int ParseMonth(string? value, string notableDateId, string ruleId, ICollection<NotableDateValidationDiagnostic> diagnostics)
-    {
-        if (value is not null)
-        {
-            if (int.TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out var numeric) && numeric is >= 1 and <= 12)
-                return numeric;
-
-            var index = Array.FindIndex(s_monthNames, n => string.Equals(n, value, StringComparison.OrdinalIgnoreCase));
-            if (index >= 0)
-                return index + 1;
-        }
-
-        diagnostics.Add(new NotableDateValidationDiagnostic(
-            NotableDateValidationSeverity.Error,
-            "BODU-CAL-MONTH",
-            string.Format(CultureInfo.InvariantCulture, CalendarResourceStrings.Validation_InvalidMonthValue, notableDateId, ruleId, value ?? string.Empty)));
-
-        return 1;
-    }
-
-    /// <summary>
-    /// Parses a fixed-strategy month for a given calendar system, returning either a numeric month or a Hebrew alias.
-    /// </summary>
-    /// <param name="value">The raw month value.</param>
-    /// <param name="calendar">The calendar system the month is expressed in.</param>
-    /// <param name="notableDateId">The identifier of the owning concept, used in diagnostics.</param>
-    /// <param name="ruleId">The identifier of the owning rule, used in diagnostics.</param>
-    /// <param name="diagnostics">The collection that receives semantic diagnostics.</param>
-    /// <returns>
-    /// A tuple of the one-based month (or <c>0</c> when an alias supplies it) and an optional Hebrew alias.
-    /// </returns>
-    private static (int Month, string? Alias) ParseFixedMonth(string? value, CalendarSystem calendar, string notableDateId, string ruleId, ICollection<NotableDateValidationDiagnostic> diagnostics)
-    {
-        if (calendar == CalendarSystem.Gregorian)
-            return (ParseMonth(value, notableDateId, ruleId, diagnostics), null);
-
-        if (value is not null)
-        {
-            if (int.TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out var numeric) && numeric is >= 1 and <= 13)
-                return (numeric, null);
-
-            switch (value)
-            {
-                case "Tishri": return (1, null);
-                case "Heshvan": return (2, null);
-                case "Kislev": return (3, null);
-                case "Tevet": return (4, null);
-                case "Shevat": return (5, null);
-                case "AdarI": return (6, null);
-                case "AdarII":
-                case "LastAdar":
-                case "Nisan":
-                case "Iyar":
-                case "Sivan":
-                case "Tammuz":
-                case "Av":
-                case "Elul":
-                    return (0, value);
-                default:
-                    break;
-            }
-        }
-
-        diagnostics.Add(new NotableDateValidationDiagnostic(
-            NotableDateValidationSeverity.Error,
-            "BODU-CAL-MONTH",
-            string.Format(CultureInfo.InvariantCulture, CalendarResourceStrings.Validation_InvalidMonthValue, notableDateId, ruleId, value ?? string.Empty)));
-
-        return (1, null);
-    }
-
-    /// <summary>
     /// Returns a child property of an object element, or <see langword="null" /> when absent or null-valued.
     /// </summary>
     /// <param name="element">The parent element.</param>
@@ -639,25 +526,4 @@ internal static class NotableDateJsonDocumentParser
 
         return values;
     }
-
-    /// <summary>
-    /// Parses an enumeration value case-insensitively, falling back to a default when absent or unrecognised.
-    /// </summary>
-    /// <typeparam name="TEnum">The enumeration type.</typeparam>
-    /// <param name="value">The raw value.</param>
-    /// <param name="fallback">The value returned when parsing fails.</param>
-    /// <returns>The parsed value, or <paramref name="fallback" />.</returns>
-    private static TEnum ParseEnum<TEnum>(string? value, TEnum fallback)
-        where TEnum : struct, Enum =>
-        value is not null && Enum.TryParse(value, ignoreCase: true, out TEnum result) && Enum.IsDefined(result) ? result : fallback;
-
-    /// <summary>
-    /// Parses an optional enumeration value case-insensitively.
-    /// </summary>
-    /// <typeparam name="TEnum">The enumeration type.</typeparam>
-    /// <param name="value">The raw value, or <see langword="null" />.</param>
-    /// <returns>The parsed value, or <see langword="null" /> when absent or unrecognised.</returns>
-    private static TEnum? ParseNullableEnum<TEnum>(string? value)
-        where TEnum : struct, Enum =>
-        value is not null && Enum.TryParse(value, ignoreCase: true, out TEnum result) && Enum.IsDefined(result) ? result : null;
 }
